refactor: colocate quest-card tests
