+++ components/quests/quest-card/index.gm.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import QuestCard from "../index";
+import QuestCard from "./index";
 import { createMockQuest } from "./quest-card.fixtures";
 import { QuestStatus } from "@/lib/types/database";
