+++ components/quests/quest-card/index.hero.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import QuestCard from "../index";
+import QuestCard from "./index";
 import { createMockQuest } from "./quest-card.fixtures";
 
 jest.mock("framer-motion", () => ({
