+++ components/quests/quest-card/quest-card-helpers.abandonment.test.ts
