+++ components/quests/quest-card/quest-card.fixtures.tsx
